% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newest_file.R
\name{newest_file}
\alias{newest_file}
\title{Get the newest file from a directory}
\usage{
newest_file(dir, pattern = NULL)
}
\arguments{
\item{dir}{A character string specifying the directory to list the files from.}

\item{pattern}{An optional regular expression. Only file names which match the
regular expression will be returned. Defaults to NULL.}
}
\value{
A character string with the path to the newest file that matches the
specified pattern.
}
\description{
This function lists all files in a given directory and returns the newest file
that matches a specified pattern. The date is extracted from the filename and
used to determine the newest file.
}
