% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save a Plot with a Timestamped Filename}
\usage{
save_plot(
  plot = ggplot2::last_plot(),
  dir,
  name = "plot",
  extension = "png",
  ...
)
}
\arguments{
\item{plot}{A plot object to save. Default is the last plot created.}

\item{dir}{A character string specifying the directory to save the plot.}

\item{name}{A character string to name the plot file. Default is "plot".}

\item{extension}{A character string for the file extension. Default is "png".}

\item{...}{Additional arguments passed to \code{ggsave}.}
}
\value{
A character string of the path to the saved plot file.
}
\description{
This function saves a plot to a specified directory with a filename that includes the current date.
If no plot is specified, the last plot created is saved. The user can also provide a name for the plot.
}
\examples{
\dontrun{
library(ggplot2)
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
save_plot(p, dir = "plots", name = "scatterplot", width = 5, height = 5)
}

}
